created fact and greeting modules
