+++ lib/app.js
@@ -1,9 +1,7 @@
 const url = require('url');
-const factArray = [
-    'REST (Representational State Transfer) was introduced and defined in 2000 by Roy Fielding in his doctoral dissertation.',
-    'Postman has an intuitive user interface to send requests, save responses, add tests, and create workflows.',
-    'URLs have a simple structure that consists of the following components: protocol, host, port, resource path, query.'
-];
+const greeting = require('./greeting');
+const fact = require('./fact');
+const cowsay = require('cowsay');
 
 function app(req, res) {
 
@@ -13,29 +11,30 @@ function app(req, res) {
 
     const path = parsedUrl.pathname.split('/')[1] || null;
 
+    let statement = '';
+
     if (req.method === 'GET') {
 
         if (path === 'greeting') {
 
-            const msg = parsedUrl.query.salutation || 'Hello';
-            const name = parsedUrl.pathname.split('/')[2] || 'stranger';
-            res.end(`${msg} ${name}`);
+            statement = greeting(parsedUrl);
 
         } else if (path === 'fact') {
 
-            const fact = factArray[Math.floor(Math.random() * factArray.length)];
-            res.end(fact);
+            statement = fact();
 
         } else {
 
-            res.end(`cannot GET ${path}`);
+            statement = `cannot GET ${path}`;
 
         }
     } else {
 
-        res.end(`cannot ${req.method} ${path}`);
+        statement = `cannot ${req.method} ${path}`;
 
     }
+
+    res.end(cowsay.say({text: statement}));
 }
 
 module.exports = app;
