+++ lib/greeting.js
@@ -1,0 +1,7 @@
+function greeting(parsedUrl) {
+    const msg = parsedUrl.query.salutation || 'Hello';
+    const name = parsedUrl.pathname.split('/')[2] || 'stranger';
+    return `${msg} ${name}`;
+}
+
+module.exports = greeting;
