+++ package.json
@@ -20,6 +20,7 @@
   "devDependencies": {
     "chai": "^4.1.0",
     "chai-http": "^3.0.0",
+    "cowsay": "^1.1.9",
     "eslint": "^4.2.0",
     "mocha": "^3.4.2"
   }
